+++ src/constants.ts
@@ -1,0 +1,3 @@
+export const NAME = 'boto3 IntelliSense';
+export const INITIALIZED = 'initialized';
+export const INSTALLER = 'installer';