+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to the "boto3-ide" extension will be documented in this file.
 
+## [Unreleased]
+### Added
+- Quick Start suggestion in a new workspace if boto3 is in use
+
+### Changed
+- If there are no packages to update, `Auto-discover` is suggested
+- Update command shows notification when update is complete
+
+### Fixed
+- Pluralized services/packages in notifications
+
 ## [0.3.0]
 ### Added
 - New `Select installer` command